+++ crates/kct_package/tests/helpers/mod.rs
@@ -1,12 +1,12 @@
 use serde_json::Value;
 use tera::{Context, Tera};
 
-pub fn values(contents: &str) -> Value {
+pub fn json(contents: &str) -> Value {
 	serde_json::from_str(contents).unwrap()
 }
 
-pub fn template(contents: &str, values: &Value) -> String {
-	let context = Context::from_serialize(values).unwrap();
+pub fn template(contents: &str, input: &Value) -> String {
+	let context = Context::from_serialize(input).unwrap();
 
 	Tera::one_off(&contents, &context, true).unwrap()
 }